Add files via upload
This project involves a web scraping application designed to collect and display COVID-19 data on a user-friendly webpage. The data includes the population of various countries, the number of reported COVID-19 cases, and the corresponding death rates. The webpage features an integrated sorting functionality, allowing users to sort data columns in ascending or descending order by simply clicking the column headers.